+++ src/main/java/com/commercetools/sync/services/impl/UnresolvedReferencesServiceImpl.java
@@ -11,27 +11,38 @@ import io.sphere.sdk.customobjects.queries.CustomObjectQuery;
 import io.sphere.sdk.queries.QueryExecutionUtils;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
+import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
+import static org.apache.commons.codec.digest.DigestUtils.sha1Hex;
 
 public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesService {
 
     private final ProductSyncOptions syncOptions;
 
-    private static final String SAVE_FAILED = "Failed to save CustomObject with key: '%s'. Reason: %s";
-    private static final String DELETE_FAILED = "Failed to delete CustomObject with key: '%s'. Reason: %s";
+    private static final String SAVE_FAILED =
+        "Failed to save CustomObject with key: '%s' (hash of product key: '%s').";
+    private static final String DELETE_FAILED =
+        "Failed to delete CustomObject with key: '%s' (hash of product key: '%s').";
     private static final String CUSTOM_OBJECT_CONTAINER_KEY =
         "commercetools-sync-java.UnresolvedReferencesService.productDrafts";
 
+
     public UnresolvedReferencesServiceImpl(@Nonnull final ProductSyncOptions baseSyncOptions) {
         this.syncOptions = baseSyncOptions;
+    }
+
+    @Nonnull
+    private String hash(@Nullable final String customObjectKey) {
+        return sha1Hex(customObjectKey);
     }
 
     @Nonnull
@@ -42,11 +53,14 @@ public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesServ
             return CompletableFuture.completedFuture(Collections.emptySet());
         }
 
+        Set<String> hashedKeys = keys.stream()
+            .map(this::hash).collect(Collectors.toSet());
+
         final CustomObjectQuery<WaitingToBeResolved> customObjectQuery =
             CustomObjectQuery
                 .of(WaitingToBeResolved.class)
                 .byContainer(CUSTOM_OBJECT_CONTAINER_KEY)
-                .plusPredicates(p -> p.key().isIn(keys));
+                .plusPredicates(p -> p.key().isIn(hashedKeys));
 
         return QueryExecutionUtils
             .queryAll(syncOptions.getCtpClient(), customObjectQuery)
@@ -64,7 +78,7 @@ public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesServ
         final CustomObjectDraft<WaitingToBeResolved> customObjectDraft = CustomObjectDraft
             .ofUnversionedUpsert(
                 CUSTOM_OBJECT_CONTAINER_KEY,
-                draft.getProductDraft().getKey(),
+                hash(draft.getProductDraft().getKey()),
                 draft,
                 WaitingToBeResolved.class);
 
@@ -76,7 +90,7 @@ public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesServ
                     return Optional.of(resource.getValue());
                 } else {
                     syncOptions.applyErrorCallback(
-                        format(SAVE_FAILED, customObjectDraft.getKey(), exception.getMessage()), exception);
+                        format(SAVE_FAILED, customObjectDraft.getKey(), draft.getProductDraft().getKey()), exception);
                     return Optional.empty();
                 }
             });
@@ -89,13 +103,13 @@ public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesServ
         return syncOptions
             .getCtpClient()
             .execute(CustomObjectDeleteCommand
-                .of(CUSTOM_OBJECT_CONTAINER_KEY, key, WaitingToBeResolved.class))
+                .of(CUSTOM_OBJECT_CONTAINER_KEY, hash(key), WaitingToBeResolved.class))
             .handle((resource, exception) -> {
                 if (exception == null) {
                     return Optional.of(resource.getValue());
                 } else {
                     syncOptions.applyErrorCallback(
-                        format(DELETE_FAILED, key, exception.getMessage()), exception);
+                        format(DELETE_FAILED, hash(key), key), exception);
                     return Optional.empty();
                 }
             });